Simplified auth flow
