+++ packages/cdk/resources/api/oauth2Functions.ts
@@ -9,10 +9,8 @@ import {SharedSecrets} from "../SharedSecrets"
 export interface OAuth2FunctionsProps {
   readonly serviceName: string
   readonly stackName: string
-  readonly fullCloudfrontDomain: string
 
   readonly fullCognitoDomain: string
-  readonly userPoolClientId: string
   readonly primaryPoolIdentityProviderName: string
   readonly mockPoolIdentityProviderName: string
 
@@ -37,16 +35,6 @@ export interface OAuth2FunctionsProps {
   readonly tokenMappingTableReadPolicy: IManagedPolicy
   readonly useTokensMappingKmsKeyPolicy: IManagedPolicy
 
-  readonly stateMappingTable: ITableV2
-  readonly stateMappingTableWritePolicy: IManagedPolicy
-  readonly stateMappingTableReadPolicy: IManagedPolicy
-  readonly useStateMappingKmsKeyPolicy: IManagedPolicy
-
-  readonly sessionStateMappingTable: ITableV2
-  readonly sessionStateMappingTableWritePolicy: IManagedPolicy
-  readonly sessionStateMappingTableReadPolicy: IManagedPolicy
-  readonly useSessionStateMappingKmsKeyPolicy: IManagedPolicy
-
   readonly sharedSecrets: SharedSecrets
 
   readonly logRetentionInDays: number
@@ -63,27 +51,11 @@ export class OAuth2Functions extends Construct {
   public readonly oAuth2Policies: Array<IManagedPolicy>
   public readonly primaryJwtPrivateKey: Secret
   public readonly authorizeLambda: NodejsFunction
-  public readonly mockAuthorizeLambda: NodejsFunction
-  public readonly callbackLambda: NodejsFunction
-  public readonly mockCallbackLambda: NodejsFunction
   public readonly tokenLambda: NodejsFunction
   public readonly mockTokenLambda: NodejsFunction
 
   public constructor(scope: Construct, id: string, props: OAuth2FunctionsProps) {
     super(scope, id)
-
-    let mockOidcAuthorizeEndpoint
-    let mockOidcIssuer
-    let mockOidcClientId
-    if (props.useMockOidc) {
-      mockOidcIssuer = props.mockOidcIssuer as string
-      mockOidcAuthorizeEndpoint = props.mockOidcAuthorizeEndpoint as string
-      mockOidcClientId = props.mockOidcClientId as string
-    } else {
-      mockOidcAuthorizeEndpoint = ""
-      mockOidcIssuer = ""
-      mockOidcClientId = ""
-    }
 
     // Create the token Lambda function
     const tokenLambda = new LambdaFunction(this, "TokenResources", {
@@ -106,10 +78,10 @@ export class OAuth2Functions extends Construct {
         CIS2_IDP_TOKEN_PATH: props.primaryOidcTokenEndpoint,
         CIS2_USER_POOL_IDP: props.primaryPoolIdentityProviderName,
         CIS2_OIDCJWKS_ENDPOINT: props.primaryOidcjwksEndpoint,
-        jwtPrivateKeyArn: props.sharedSecrets.primaryJwtPrivateKey.secretArn,
         CIS2_OIDC_CLIENT_ID: props.primaryOidcClientId,
         CIS2_OIDC_ISSUER: props.primaryOidcIssuer,
-        FULL_CLOUDFRONT_DOMAIN: props.fullCloudfrontDomain,
+        COGNITO_DOMAIN: props.fullCognitoDomain,
+        jwtPrivateKeyArn: props.sharedSecrets.primaryJwtPrivateKey.secretArn,
         jwtKid: props.jwtKid
       }
     })
@@ -119,55 +91,21 @@ export class OAuth2Functions extends Construct {
       serviceName: props.serviceName,
       stackName: props.stackName,
       lambdaName: `${props.stackName}-authorize`,
-      additionalPolicies: [
-        props.stateMappingTableWritePolicy,
-        props.stateMappingTableReadPolicy,
-        props.useStateMappingKmsKeyPolicy
-      ],
       logRetentionInDays: props.logRetentionInDays,
       logLevel: props.logLevel,
       packageBasePath: "packages/cognito",
       entryPoint: "src/authorize.ts",
       lambdaEnvironmentVariables: {
-        IDP_AUTHORIZE_PATH: props.primaryOidcAuthorizeEndpoint,
-        OIDC_CLIENT_ID: props.primaryOidcClientId,
-        COGNITO_CLIENT_ID: props.userPoolClientId,
-        FULL_CLOUDFRONT_DOMAIN: props.fullCloudfrontDomain,
-        StateMappingTableName: props.stateMappingTable.tableName
-      }
-    })
-
-    // This proxy handles the return journey from the IdP login initiated by the authorize lambda
-    const callbackLambda = new LambdaFunction(this, "CallbackLambdaResources", {
-      serviceName: props.serviceName,
-      stackName: props.stackName,
-      lambdaName: `${props.stackName}-callback`,
-      additionalPolicies: [
-        props.stateMappingTableWritePolicy,
-        props.stateMappingTableReadPolicy,
-        props.useStateMappingKmsKeyPolicy
-      ],
-      logRetentionInDays: props.logRetentionInDays,
-      logLevel: props.logLevel,
-      packageBasePath: "packages/cognito",
-      entryPoint: "src/callback.ts",
-      lambdaEnvironmentVariables: {
-        StateMappingTableName: props.stateMappingTable.tableName,
-        COGNITO_CLIENT_ID: props.userPoolClientId,
-        COGNITO_DOMAIN: props.fullCognitoDomain,
-        MOCK_OIDC_ISSUER: mockOidcIssuer,
-        PRIMARY_OIDC_ISSUER: props.primaryOidcIssuer
+        IDP_AUTHORIZE_PATH: props.primaryOidcAuthorizeEndpoint
       }
     })
 
     // Initialize policies
     const oauth2Policies: Array<IManagedPolicy> = [
       authorizeLambda.executeLambdaManagedPolicy,
-      callbackLambda.executeLambdaManagedPolicy,
       tokenLambda.executeLambdaManagedPolicy
     ]
 
-    let mockAuthorizeLambda: LambdaFunction
     let mockTokenLambda: LambdaFunction
     if (props.useMockOidc) {
       if (
@@ -180,37 +118,6 @@ export class OAuth2Functions extends Construct {
         throw new Error("Missing mock OIDC configuration.")
       }
 
-      mockAuthorizeLambda = new LambdaFunction(this, "MockAuthorizeLambdaResources", {
-        serviceName: props.serviceName,
-        stackName: props.stackName,
-        lambdaName: `${props.stackName}-mock-authorize`,
-        additionalPolicies: [
-          props.stateMappingTableWritePolicy,
-          props.stateMappingTableReadPolicy,
-          props.useStateMappingKmsKeyPolicy,
-          props.sessionStateMappingTableWritePolicy,
-          props.sessionStateMappingTableReadPolicy,
-          props.useSessionStateMappingKmsKeyPolicy
-        ],
-        logRetentionInDays: props.logRetentionInDays,
-        logLevel: props.logLevel,
-        packageBasePath: "packages/cognito",
-        entryPoint: "src/authorizeMock.ts",
-        lambdaEnvironmentVariables: {
-          IDP_AUTHORIZE_PATH: mockOidcAuthorizeEndpoint,
-          OIDC_CLIENT_ID: mockOidcClientId,
-          COGNITO_CLIENT_ID: props.userPoolClientId,
-          FULL_CLOUDFRONT_DOMAIN: props.fullCloudfrontDomain,
-          StateMappingTableName: props.stateMappingTable.tableName,
-          SessionStateMappingTableName: props.sessionStateMappingTable.tableName,
-          APIGEE_API_KEY: props.apigeeApiKey
-        }
-      })
-
-      oauth2Policies.push(
-        mockAuthorizeLambda.executeLambdaManagedPolicy
-      )
-
       mockTokenLambda = new LambdaFunction(this, "MockTokenResources", {
         serviceName: props.serviceName,
         stackName: props.stackName,
@@ -218,15 +125,7 @@ export class OAuth2Functions extends Construct {
         additionalPolicies: [
           props.tokenMappingTableWritePolicy,
           props.tokenMappingTableReadPolicy,
-          props.useTokensMappingKmsKeyPolicy,
-          props.stateMappingTableReadPolicy,
-          props.stateMappingTableWritePolicy,
-          props.useStateMappingKmsKeyPolicy,
-          props.sessionStateMappingTableReadPolicy,
-          props.sessionStateMappingTableWritePolicy,
-          props.useSessionStateMappingKmsKeyPolicy,
-          props.sharedSecrets.useJwtKmsKeyPolicy,
-          props.sharedSecrets.getMockJwtPrivateKeyPolicy
+          props.useTokensMappingKmsKeyPolicy
         ],
         logRetentionInDays: props.logRetentionInDays,
         logLevel: props.logLevel,
@@ -234,18 +133,13 @@ export class OAuth2Functions extends Construct {
         entryPoint: "src/tokenMock.ts",
         lambdaEnvironmentVariables: {
           TokenMappingTableName: props.tokenMappingTable.tableName,
-          SessionStateMappingTableName: props.sessionStateMappingTable.tableName,
-          StateMappingTableName: props.stateMappingTable.tableName,
-          MOCK_IDP_TOKEN_PATH: props.mockOidcTokenEndpoint,
           MOCK_USER_POOL_IDP: props.mockPoolIdentityProviderName,
           MOCK_OIDCJWKS_ENDPOINT: props.mockOidcjwksEndpoint,
-          jwtPrivateKeyArn: props.sharedSecrets.mockJwtPrivateKey!.secretArn,
           MOCK_OIDC_CLIENT_ID: props.mockOidcClientId,
           MOCK_OIDC_TOKEN_ENDPOINT: props.mockOidcTokenEndpoint,
           MOCK_USER_INFO_ENDPOINT: props.mockOidcUserInfoEndpoint,
           MOCK_OIDC_ISSUER: props.mockOidcIssuer,
-          FULL_CLOUDFRONT_DOMAIN: props.fullCloudfrontDomain,
-          jwtKid: props.jwtKid,
+          COGNITO_DOMAIN: props.fullCognitoDomain,
           APIGEE_API_KEY: props.apigeeApiKey,
           APIGEE_API_SECRET: props.apigeeApiSecret
         }
@@ -255,40 +149,8 @@ export class OAuth2Functions extends Construct {
         mockTokenLambda.executeLambdaManagedPolicy
       )
 
-      const mockCallbackLambda = new LambdaFunction(this, "MockCallbackLambdaResources", {
-        serviceName: props.serviceName,
-        stackName: props.stackName,
-        lambdaName: `${props.stackName}-mock-callback`,
-        additionalPolicies: [
-          props.stateMappingTableWritePolicy,
-          props.stateMappingTableReadPolicy,
-          props.useStateMappingKmsKeyPolicy,
-          props.sessionStateMappingTableReadPolicy,
-          props.sessionStateMappingTableWritePolicy,
-          props.useSessionStateMappingKmsKeyPolicy
-        ],
-        logRetentionInDays: props.logRetentionInDays,
-        logLevel: props.logLevel,
-        packageBasePath: "packages/cognito",
-        entryPoint: "src/callbackMock.ts",
-        lambdaEnvironmentVariables: {
-          StateMappingTableName: props.stateMappingTable.tableName,
-          SessionStateMappingTableName: props.sessionStateMappingTable.tableName,
-          COGNITO_CLIENT_ID: props.userPoolClientId,
-          COGNITO_DOMAIN: props.fullCognitoDomain,
-          MOCK_OIDC_ISSUER: mockOidcIssuer,
-          PRIMARY_OIDC_ISSUER: props.primaryOidcIssuer
-        }
-      })
-
-      oauth2Policies.push(
-        mockCallbackLambda.executeLambdaManagedPolicy
-      )
-
       // Output
-      this.mockAuthorizeLambda = mockAuthorizeLambda.lambda
       this.mockTokenLambda = mockTokenLambda.lambda
-      this.mockCallbackLambda = mockCallbackLambda.lambda
     }
 
     // Outputs
@@ -296,7 +158,6 @@ export class OAuth2Functions extends Construct {
     this.primaryJwtPrivateKey = props.sharedSecrets.primaryJwtPrivateKey
 
     this.authorizeLambda = authorizeLambda.lambda
-    this.callbackLambda = callbackLambda.lambda
     this.tokenLambda = tokenLambda.lambda
   }
 }
