+++ packages/cognito/src/token.ts
@@ -23,29 +23,19 @@ It expects the following environment variables to be set
 TokenMappingTableName
 jwtPrivateKeyArn
 jwtKid
+COGNITO_DOMAIN
 
-For CIS2 calls, the following must be set
 CIS2_OIDC_ISSUER
 CIS2_OIDC_CLIENT_ID
 CIS2_OIDCJWKS_ENDPOINT
 CIS2_USER_INFO_ENDPOINT
 CIS2_USER_POOL_IDP
 CIS2_IDP_TOKEN_PATH
-FULL_CLOUDFRONT_DOMAIN
-
-For mock calls, the following must be set
-MOCK_OIDC_ISSUER
-MOCK_OIDC_CLIENT_ID
-MOCK_OIDCJWKS_ENDPOINT
-MOCK_USER_INFO_ENDPOINT
-MOCK_USER_POOL_IDP
-MOCK_IDP_TOKEN_PATH
-FULL_CLOUDFRONT_DOMAIN
 */
 
 const logger = new Logger({serviceName: "token"})
 
-const cloudfrontDomain = process.env["FULL_CLOUDFRONT_DOMAIN"] as string
+const cognitoDomain = process.env["COGNITO_DOMAIN"] as string
 
 // Create a config for cis2 and mock
 // this is outside functions so it can be re-used and caching works
@@ -55,12 +45,10 @@ const TokenMappingTableName = process.env["TokenMappingTableName"] as string
 const jwtPrivateKeyArn = process.env["jwtPrivateKeyArn"] as string
 const jwtKid = process.env["jwtKid"] as string
 
-const SessionStateMappingTableName = process.env["SessionStateMappingTableName"] as string
-
 const idpTokenPath = process.env["CIS2_IDP_TOKEN_PATH"] as string
 
 // Set the redirect back to our proxy lambda
-const idpCallbackPath = `https://${cloudfrontDomain}/oauth2/callback`
+const idpCallbackPath = `https://${cognitoDomain}/oauth2/idpresponse`
 
 const dynamoClient = new DynamoDBClient()
 const documentClient = DynamoDBDocumentClient.from(dynamoClient)
@@ -77,11 +65,10 @@ const lambdaHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayPro
   })
   const axiosInstance = axios.create()
   logger.debug("data from env variables", {
-    cloudfrontDomain,
+    cognitoDomain,
     TokenMappingTableName,
     jwtPrivateKeyArn,
     jwtKid,
-    SessionStateMappingTableName,
     idpTokenPath
   })
