--- packages/trackerUserInfoLambda/tests/test_handler.mockEnabled.test.ts
@@ -1,440 +1,0 @@
-import {jest} from "@jest/globals"
-import jwksClient from "jwks-rsa"
-import {OidcConfig} from "@cpt-ui-common/authFunctions"
-import {TrackerUserInfo} from "@/userInfoTypes"
-
-process.env.MOCK_MODE_ENABLED = "true"
-
-// Mocked functions from authFunctions
-const mockFetchAndVerifyCIS2Tokens = jest.fn()
-const mockGetUsernameFromEvent = jest.fn()
-const mockInitializeOidcConfig = jest.fn()
-const mockAuthenticateRequest = jest.fn()
-
-jest.unstable_mockModule("@cpt-ui-common/authFunctions", () => {
-  const fetchAndVerifyCIS2Tokens = mockFetchAndVerifyCIS2Tokens.mockImplementation(async () => {
-    return {
-      cis2IdToken: "idToken",
-      cis2AccessToken: "accessToken"
-    }
-  })
-
-  const getUsernameFromEvent = mockGetUsernameFromEvent.mockImplementation(() => {
-    return "Mock_JoeBloggs"
-  })
-
-  const authenticateRequest = mockAuthenticateRequest.mockImplementation(async (event) => {
-    const username = mockGetUsernameFromEvent(event)
-    return {
-      username,
-      apigeeAccessToken: "foo",
-      cis2IdToken: "mock-id-token",
-      roleId: "test-role",
-      isMockRequest: typeof username === "string" && username.startsWith("Mock_")
-    }
-  })
-
-  const initializeOidcConfig = mockInitializeOidcConfig.mockImplementation(() => {
-    // Create a JWKS client for cis2 and mock
-    const cis2JwksUri = process.env["CIS2_OIDCJWKS_ENDPOINT"] as string
-    const cis2JwksClient = jwksClient({
-      jwksUri: cis2JwksUri,
-      cache: true,
-      cacheMaxEntries: 5,
-      cacheMaxAge: 3600000 // 1 hour
-    })
-
-    const cis2OidcConfig: OidcConfig = {
-      oidcIssuer: process.env["CIS2_OIDC_ISSUER"] ?? "",
-      oidcClientID: process.env["CIS2_OIDC_CLIENT_ID"] ?? "",
-      oidcJwksEndpoint: process.env["CIS2_OIDCJWKS_ENDPOINT"] ?? "",
-      oidcUserInfoEndpoint: process.env["CIS2_USER_INFO_ENDPOINT"] ?? "",
-      userPoolIdp: process.env["CIS2_USER_POOL_IDP"] ?? "",
-      oidcTokenEndpoint: process.env["CIS2_IDP_TOKEN_PATH"] ?? "",
-      jwksClient: cis2JwksClient,
-      tokenMappingTableName: process.env["TokenMappingTableName"] ?? ""
-    }
-
-    const mockJwksUri = process.env["MOCK_OIDCJWKS_ENDPOINT"] as string
-    const mockJwksClient = jwksClient({
-      jwksUri: mockJwksUri,
-      cache: true,
-      cacheMaxEntries: 5,
-      cacheMaxAge: 3600000 // 1 hour
-    })
-
-    const mockOidcConfig: OidcConfig = {
-      oidcIssuer: process.env["MOCK_OIDC_ISSUER"] ?? "",
-      oidcClientID: process.env["MOCK_OIDC_CLIENT_ID"] ?? "",
-      oidcJwksEndpoint: process.env["MOCK_OIDCJWKS_ENDPOINT"] ?? "",
-      oidcUserInfoEndpoint: process.env["MOCK_USER_INFO_ENDPOINT"] ?? "",
-      userPoolIdp: process.env["MOCK_USER_POOL_IDP"] ?? "",
-      oidcTokenEndpoint: process.env["MOCK_IDP_TOKEN_PATH"] ?? "",
-      jwksClient: mockJwksClient,
-      tokenMappingTableName: process.env["TokenMappingTableName"] ?? ""
-    }
-
-    return {cis2OidcConfig, mockOidcConfig}
-  })
-
-  return {
-    fetchAndVerifyCIS2Tokens,
-    getUsernameFromEvent,
-    initializeOidcConfig,
-    authenticateRequest
-  }
-})
-
-// Mocked functions from userInfoHelpers
-const mockFetchUserInfo = jest.fn()
-const mockUpdateDynamoTable = jest.fn()
-const mockFetchDynamoTable = jest.fn()
-
-jest.unstable_mockModule("@/userInfoHelpers", () => {
-  const fetchUserInfo = mockFetchUserInfo.mockImplementation(() => {
-    return {
-      roles_with_access: [],
-      roles_without_access: [],
-      selected_role: {}
-    }
-  })
-
-  const updateDynamoTable = mockUpdateDynamoTable.mockImplementation(() => {})
-
-  const fetchDynamoTable = mockFetchDynamoTable.mockImplementation(() => {
-    return {
-      roles_with_access: [],
-      roles_without_access: [],
-      currently_selected_role: {}
-    }
-  })
-
-  return {
-    fetchUserInfo,
-    updateDynamoTable,
-    fetchDynamoTable
-  }
-})
-
-const {handler} = await import("@/handler")
-import {mockContext, mockAPIGatewayProxyEvent} from "./mockObjects"
-import {Logger} from "@aws-lambda-powertools/logger"
-
-describe("Lambda Handler Tests with mock enabled", () => {
-  let event = {...mockAPIGatewayProxyEvent}
-  let context = {...mockContext}
-
-  beforeEach(() => {
-    jest.clearAllMocks()
-
-    // Reset all mocks to their default implementation for each test
-    mockFetchUserInfo.mockImplementation(() => {
-      return {
-        roles_with_access: [],
-        roles_without_access: [],
-        selected_role: {}
-      }
-    })
-    mockUpdateDynamoTable.mockImplementation(() => {})
-    mockFetchAndVerifyCIS2Tokens.mockImplementation(async () => {
-      return {
-        cis2IdToken: "idToken",
-        cis2AccessToken: "accessToken"
-      }
-    })
-    mockAuthenticateRequest.mockImplementation(async (event) => {
-      const username = mockGetUsernameFromEvent(event)
-      return {
-        username,
-        apigeeAccessToken: "foo",
-        cis2IdToken: "mock-id-token",
-        roleId: "test-role",
-        isMockRequest: typeof username === "string" && username.startsWith("Mock_")
-      }
-    })
-    mockGetUsernameFromEvent.mockImplementation(() => {
-      return "Mock_JoeBloggs"
-    })
-  })
-
-  it("should return a successful response when called normally", async () => {
-    const response = await handler(event, context)
-
-    expect(mockAuthenticateRequest).toHaveBeenCalled()
-    expect(mockFetchUserInfo).toHaveBeenCalled()
-    expect(mockUpdateDynamoTable).toHaveBeenCalled()
-
-    expect(response).toBeDefined()
-    expect(response).toHaveProperty("statusCode", 200)
-    expect(response).toHaveProperty("body")
-
-    const body = JSON.parse(response.body)
-    expect(body).toHaveProperty("message", "UserInfo fetched successfully from the OIDC endpoint")
-    expect(body).toHaveProperty("userInfo")
-  })
-
-  it("should use cis2 values when username does not start with Mock_", async () => {
-    mockGetUsernameFromEvent.mockReturnValue("Primary_JohnDoe")
-
-    // Update the expectation to match what's actually passed
-    await handler(event, context)
-    expect(mockGetUsernameFromEvent).toHaveBeenCalled()
-
-    // Check that authenticateRequest was called, but don't be specific about exact parameters
-    expect(mockAuthenticateRequest).toHaveBeenCalled()
-
-    // For fetchUserInfo, check if it was called with the right OidcConfig type parameters
-    expect(mockFetchUserInfo).toHaveBeenCalledWith(
-      expect.any(String),
-      expect.any(String),
-      expect.any(Object),
-      expect.any(Object),
-      expect.objectContaining({
-        oidcIssuer: expect.any(String),
-        oidcClientID: expect.any(String),
-        oidcJwksEndpoint: expect.any(String),
-        oidcUserInfoEndpoint: expect.any(String),
-        userPoolIdp: expect.any(String),
-        tokenMappingTableName: expect.any(String)
-      })
-    )
-  })
-
-  it("should use mock values when username starts with Mock_", async () => {
-    mockGetUsernameFromEvent.mockReturnValue("Mock_JaneDoe")
-
-    await handler(event, context)
-
-    // Check that authenticateRequest was called, but don't be specific about exact parameters
-    expect(mockAuthenticateRequest).toHaveBeenCalled()
-
-    // Check that fetchUserInfo was called with expected parameters (loosely)
-    expect(mockFetchUserInfo).toHaveBeenCalledWith(
-      expect.any(String),
-      expect.any(String),
-      expect.any(Object),
-      expect.any(Object),
-      expect.objectContaining({
-        oidcIssuer: expect.any(String),
-        oidcClientID: expect.any(String),
-        oidcJwksEndpoint: expect.any(String),
-        oidcUserInfoEndpoint: expect.any(String),
-        userPoolIdp: expect.any(String),
-        tokenMappingTableName: expect.any(String)
-      })
-    )
-  })
-
-  it("should return error when authenticateRequest throws an error", async () => {
-    const error = new Error("Token verification failed")
-    const loggerSpy = jest.spyOn(Logger.prototype, "error")
-
-    // Important: Use implementation, not return value
-    mockAuthenticateRequest.mockImplementationOnce(() => {
-      throw error
-    })
-
-    const response = await handler(event, context)
-
-    // Check response format matches what the middleware produces
-    expect(response).toMatchObject({
-      message: "A system error has occurred"
-    })
-
-    expect(loggerSpy).toHaveBeenCalledWith(
-      expect.any(Object),
-      "Error: Token verification failed"
-    )
-    expect(mockFetchUserInfo).not.toHaveBeenCalled()
-    expect(mockUpdateDynamoTable).not.toHaveBeenCalled()
-  })
-
-  it("should return error when fetchUserInfo throws an error", async () => {
-    const error = new Error("User info fetch failed")
-    const loggerSpy = jest.spyOn(Logger.prototype, "error")
-
-    // Use implementation to throw an error synchronously
-    mockFetchUserInfo.mockImplementationOnce(() => {
-      throw error
-    })
-
-    const response = await handler(event, context)
-    expect(response).toMatchObject({
-      message: "A system error has occurred"
-    })
-
-    expect(loggerSpy).toHaveBeenCalledWith(
-      expect.any(Object),
-      "Error: User info fetch failed"
-    )
-    expect(mockUpdateDynamoTable).not.toHaveBeenCalled()
-  })
-
-  it("should return error when updateDynamoTable throws an error", async () => {
-    const error = new Error("Dynamo update failed")
-    const loggerSpy = jest.spyOn(Logger.prototype, "error")
-
-    // Use implementation to throw an error synchronously
-    mockUpdateDynamoTable.mockImplementationOnce(() => {
-      throw error
-    })
-
-    const response = await handler(event, context)
-    expect(response).toMatchObject({
-      message: "A system error has occurred"
-    })
-
-    expect(loggerSpy).toHaveBeenCalledWith(
-      expect.any(Object),
-      "Error: Dynamo update failed"
-    )
-  })
-
-  it("should handle unexpected error types gracefully", async () => {
-    const error = new Error("Unexpected error string")
-    const loggerSpy = jest.spyOn(Logger.prototype, "error")
-
-    // Use implementation to throw an error synchronously
-    mockUpdateDynamoTable.mockImplementationOnce(() => {
-      throw error
-    })
-
-    const response = await handler(event, context)
-    expect(response).toMatchObject({
-      message: "A system error has occurred"
-    })
-
-    expect(loggerSpy).toHaveBeenCalledWith(
-      expect.any(Object),
-      "Error: Unexpected error string"
-    )
-  })
-
-  it("should call updateDynamoTable with the correct parameters", async () => {
-    const testUsername = "Mock_JaneDoe"
-    mockGetUsernameFromEvent.mockReturnValue(testUsername)
-    const userInfoMock = {
-      roles_with_access: ["roleX"],
-      roles_without_access: ["roleY"],
-      selected_role: ["roleX"]
-    }
-    mockFetchUserInfo.mockReturnValue(userInfoMock)
-
-    await handler(event, context)
-
-    // Check if updateDynamoTable was called with expected parameters
-    expect(mockUpdateDynamoTable).toHaveBeenCalled()
-    expect(mockUpdateDynamoTable.mock.calls[0][0]).toBe(testUsername)
-    expect(mockUpdateDynamoTable.mock.calls[0][1]).toEqual(userInfoMock)
-    // We don't need to be too strict about the other parameters
-  })
-
-  it("should return user info if roles_with_access is not empty", async () => {
-    const testUsername = "Mock_JaneDoe"
-    mockGetUsernameFromEvent.mockReturnValue(testUsername)
-
-    const userInfoMock: TrackerUserInfo = {
-      roles_with_access: [
-        {role_name: "Doctor", role_id: "123", org_code: "ABC", org_name: "Test Hospital"}
-      ],
-      roles_without_access: [],
-      currently_selected_role: {
-        role_name: "Doctor",
-        role_id: "123",
-        org_code: "ABC",
-        org_name: "Test Hospital"
-      },
-      user_details: {family_name: "Doe", given_name: "John"}
-    }
-
-    // Fix by explicitly casting to the expected return type
-    mockFetchDynamoTable.mockResolvedValue(userInfoMock as never)
-
-    const response = await handler(event, context)
-
-    expect(mockFetchDynamoTable).toHaveBeenCalledWith(
-      testUsername,
-      expect.any(Object), // documentClient instance
-      expect.any(Object), // logger instance
-      expect.any(String) // tokenMappingTableName
-    )
-
-    expect(response.statusCode).toBe(200)
-    expect(response.body).toContain("UserInfo fetched successfully from DynamoDB")
-
-    const responseBody = JSON.parse(response.body)
-    expect(responseBody.userInfo).toEqual(userInfoMock)
-  })
-
-  it("should return user info if roles_without_access is not empty", async () => {
-    const testUsername = "Mock_JaneDoe"
-    mockGetUsernameFromEvent.mockReturnValue(testUsername)
-
-    const userInfoMock: TrackerUserInfo = {
-      roles_with_access: [
-        {role_name: "Doctor", role_id: "123", org_code: "ABC", org_name: "Test Hospital"}
-      ],
-      roles_without_access: [],
-      currently_selected_role: {
-        role_name: "Doctor",
-        role_id: "123",
-        org_code: "ABC",
-        org_name: "Test Hospital"
-      },
-      user_details: {family_name: "Doe", given_name: "John"}
-    }
-
-    mockFetchDynamoTable.mockResolvedValue(userInfoMock as never)
-
-    const response = await handler(event, context)
-
-    expect(mockFetchDynamoTable).toHaveBeenCalledWith(
-      testUsername,
-      expect.any(Object),
-      expect.any(Object),
-      expect.any(String)
-    )
-
-    expect(response.statusCode).toBe(200)
-    expect(response.body).toContain("UserInfo fetched successfully from DynamoDB")
-
-    const responseBody = JSON.parse(response.body)
-    expect(responseBody.userInfo).toEqual(userInfoMock)
-  })
-
-  it("should return cached user info even if currently_selected_role is undefined", async () => {
-    const testUsername = "Mock_JaneDoe"
-    mockGetUsernameFromEvent.mockReturnValue(testUsername)
-
-    const userInfoMock: TrackerUserInfo = {
-      roles_with_access: [
-        {role_name: "Doctor", role_id: "123", org_code: "ABC", org_name: "Test Hospital"}
-      ],
-      roles_without_access: [],
-      currently_selected_role: {
-        role_name: "Doctor",
-        role_id: "123",
-        org_code: "ABC",
-        org_name: "Test Hospital"
-      },
-      user_details: {family_name: "Doe", given_name: "John"}
-    }
-
-    mockFetchDynamoTable.mockResolvedValue(userInfoMock as never)
-
-    const response = await handler(event, context)
-
-    expect(mockFetchDynamoTable).toHaveBeenCalledWith(
-      testUsername,
-      expect.any(Object),
-      expect.any(Object),
-      expect.any(String)
-    )
-
-    expect(response.statusCode).toBe(200)
-    expect(response.body).toContain("UserInfo fetched successfully from DynamoDB")
-
-    const responseBody = JSON.parse(response.body)
-    expect(responseBody.userInfo).toEqual(userInfoMock)
-  })
-})
